Implement generalized surface_flux FluxCalc struct
verified over land (const z0)
