+++ Source/BoundaryConditions/ERF_MOSTStress.H
@@ -6,113 +6,6 @@
 #include <ERF_Wstar.H>
 
 /**
- * Surface flux with constant roughness
- */
-struct surface_flux
-{
-    surface_flux (amrex::Real zref,
-                  amrex::Real Tflux,
-                  amrex::Real Qvflux,
-                  bool cons_qflux)
-    {
-        mdata.zref = zref;
-        mdata.surf_temp_flux  = Tflux;
-        mdata.surf_moist_flux = Qvflux;
-        spec_qflux = cons_qflux;
-    }
-
-    AMREX_GPU_DEVICE
-    AMREX_FORCE_INLINE
-    void
-    iterate_flux (const int& i,
-                  const int& j,
-                  const int& k,
-                  const int& max_iters,
-                  const amrex::Array4<amrex::Real>& z0_arr,
-                  const amrex::Array4<const amrex::Real>& umm_arr,
-                  const amrex::Array4<const amrex::Real>& tm_arr,
-                  const amrex::Array4<const amrex::Real>& tvm_arr,
-                  const amrex::Array4<const amrex::Real>& qvm_arr,
-                  const amrex::Array4<amrex::Real>& u_star_arr,
-                  const amrex::Array4<amrex::Real>& w_star_arr,
-                  const amrex::Array4<amrex::Real>& t_star_arr,
-                  const amrex::Array4<amrex::Real>& q_star_arr,
-                  const amrex::Array4<amrex::Real>& t_surf_arr,
-                  const amrex::Array4<amrex::Real>& q_surf_arr,
-                  const amrex::Array4<amrex::Real>& olen_arr,
-                  const amrex::Array4<amrex::Real>& pblh_arr,
-                  const amrex::Array4<amrex::Real>& /*Hwave_arr*/,
-                  const amrex::Array4<amrex::Real>& /*Lwave_arr*/,
-                  const amrex::Array4<amrex::Real>& /*eta_arr*/,
-                  const RoughnessModel& update_roughness) const
-    {
-        int iter = 0;
-        amrex::Real ustar = 0.0;
-        amrex::Real wstar = 0.0;
-        amrex::Real tflux = 0.0;
-        amrex::Real qflux = 0.0;
-        amrex::Real zeta  = 0.0;
-        amrex::Real psi_m = 0.0;
-        amrex::Real psi_h = 0.0;
-        amrex::Real Olen  = 0.0;
-        amrex::Real umm   = std::max(umm_arr(i,j,k), WSMIN);
-        if (u_star_arr(i,j,k) == 1.E34) {
-            u_star_arr(i,j,k) = mdata.kappa * umm / std::log(mdata.zref / z0_arr(i,j,k));
-        } else {
-            Olen  = olen_arr(i,j,k);
-            zeta  = mdata.zref / Olen;
-            psi_m = sfuns.calc_psi_m(zeta);
-            psi_h = sfuns.calc_psi_h(zeta);
-        }
-        do {
-            ustar = u_star_arr(i,j,k);
-            update_roughness(z0_arr(i,j,k), ustar, umm, tm_arr(i,j,k), 0,
-                             0, 0, 0);
-            qflux = (spec_qflux) ? mdata.surf_moist_flux :
-                                 -(qvm_arr(i,j,k) - q_surf_arr(i,j,k)) * ustar * mdata.kappa /
-                                  (std::log(mdata.zref / z0_arr(i,j,k)) - psi_h); // <w'Qv'>
-            tflux = mdata.surf_temp_flux*(1 + 0.61*qvm_arr(i,j,k)) + qflux*0.61*tm_arr(i,j,k);
-            if (w_star_arr) {
-                // update w* and Umagmean
-                w_star_arr(i,j,k) = calc_wstar(tflux, pblh_arr(i,j,k), tvm_arr(i,j,k));
-                wstar = mdata.Bjr_beta * w_star_arr(i,j,k);
-                umm = std::sqrt(umm_arr(i,j,k)*umm_arr(i,j,k) + wstar*wstar);
-                umm = std::max(umm, WSMIN);
-            }
-            Olen  = -ustar * ustar * ustar * tvm_arr(i,j,k) / (mdata.kappa * mdata.gravity * tflux);
-            zeta  = mdata.zref / Olen;
-            psi_m = sfuns.calc_psi_m(zeta);
-            psi_h = sfuns.calc_psi_h(zeta);
-            u_star_arr(i,j,k) = mdata.kappa * umm / (std::log(mdata.zref / z0_arr(i,j,k)) - psi_m);
-            ++iter;
-        } while ((std::abs(u_star_arr(i,j,k) - ustar) > tol) && iter <= max_iters);
-        AMREX_ASSERT_WITH_MESSAGE(iter < max_iters, "Maximum number of MOST iterations reached.");
-
-        t_surf_arr(i,j,k) = mdata.surf_temp_flux * (std::log(mdata.zref / z0_arr(i,j,k)) - psi_h) /
-                            (u_star_arr(i,j,k) * mdata.kappa) + tm_arr(i,j,k);
-        t_star_arr(i,j,k) = -mdata.surf_temp_flux / u_star_arr(i,j,k);
-          olen_arr(i,j,k) = Olen;
-        if (spec_qflux) {
-            q_surf_arr(i,j,k) = mdata.surf_moist_flux * (std::log(mdata.zref / z0_arr(i,j,k)) - psi_h) /
-                                (u_star_arr(i,j,k) * mdata.kappa) + qvm_arr(i,j,k);
-            q_star_arr(i,j,k) = -mdata.surf_moist_flux / u_star_arr(i,j,k);
-        } else {
-            q_star_arr(i,j,k) = mdata.kappa * (qvm_arr(i,j,k) - q_surf_arr(i,j,k)) /
-                                (std::log(mdata.zref / z0_arr(i,j,k)) - psi_h);
-        }
-    }
-
-private:
-    most_data mdata;
-    bool spec_qflux;
-    similarity_funs sfuns;
-
-    static constexpr amrex::Real tol{1.0e-5};
-    static constexpr amrex::Real WSMIN{0.1}; // minimum wind speed
-};
-
-
-/**
  * Surface flux with charnock roughness
  */
 struct surface_flux_charnock
